+++ Items/Weapons/Magic/StormsCall.cs
@@ -35,7 +35,7 @@ namespace ProvidenceMod.Items.Weapons.Magic
       item.autoReuse = true;
       item.rare = ItemRarityID.Orange;
       item.UseSound = SoundID.Item45;
-      item.Providence().element = ElementID.Air; // Fire
+      item.Providence().element = (int)ElementID.Air; // Fire
       item.shoot = ProjectileType<ZephyrSpirit>();
       item.shootSpeed = 6f;
     }
